Staging rakib

+++ src/Dashboard/DashboardPages/JoinedEventUpdatePage/JoinedEventUpdatePage.js
@@ -26,7 +26,7 @@ const JoinedEventUpdatePage = () => {
   }, [user?.email]);
 
   const handleUpdate = (id) => {
-    console.log(id);
+    // console.log(id);
 
     for (const index in Events) {
       if (Events[index]._id === id) {
@@ -80,18 +80,25 @@ const JoinedEventUpdatePage = () => {
   };
 
   const handleDelete = (id) => {
-    fetch(
-      `https://alumni-managemnet-app-server.vercel.app/join-event/delete/${id}`,
-      {
-        method: "DELETE",
-      }
-    )
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          toast.success(`Delete Successfully.`);
+
+    const agree = window.confirm(
+      `Are you want to Delete ?`
+    );
+
+    if(agree){
+      fetch(
+        `https://alumni-managemnet-app-server.vercel.app/join-event/delete/${id}`,
+        {
+          method: "DELETE",
         }
-      });
+      )
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            toast.success(`Delete Successfully.`);
+          }
+        });
+    }
   };
 
   return (
