+++ src/Pages/SingleEvent/SingleEvent.js
@@ -94,18 +94,25 @@ const SingleEvent = () => {
   };
 
   const handleDelete = () => {
-    fetch(
-      `https://alumni-managemnet-app-server.vercel.app/join-event/delete/${eventData._id}`,
-      {
-        method: "DELETE",
-      }
-    )
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          toast.success(`Delete Successfully.`);
+
+    const agree = window.confirm(
+      `Are you want to Delete ?`
+    );
+
+    if(agree){
+      fetch(
+        `https://alumni-managemnet-app-server.vercel.app/join-event/delete/${eventData._id}`,
+        {
+          method: "DELETE",
         }
-      });
+      )
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            toast.success(`Delete Successfully.`);
+          }
+        });
+    }
   };
 
   const form = useRef();
